feat: add worker selection and genserver call to worker
Implement a simple get request on _/fn/name path (default namespace) to perform selection of a worker,
with the Scheduler module, and a GenServer.call to send a prepare_container message.
It retrieves the node list, filters for the nodes with worker in the address and passes the list
to the rust module, which returns the first worker if any. It is then used to call a prepare_container.