refactor: change architecture to hexagonal (ports and adapters)
Start of a big refactor of the project. Move bandit server to driver adapter 'requests' (Core.Adapters.Requests.Http.Router).
Move the communication with the workers to a driven adapter 'commands' that the core uses to send commands to workers.
Add Mox library to create adapters mock to use in tests and update nif name to Elixir.Core.Nif.Scheduler (with rustler update to 0.25.0).