refactor: add cluster port to be used to get info from the cluster
Add a cluster port implemented in the cluster adapter.
It is used to get info about the cluster (for now just Node.list).
This way it is separated from the domain and can be mocked.
Add a Nodes module that uses the cluster port and does processing on the data,
such as filtering the nodes to get only the worker_nodes.
This was in the invoker before, so now it is moved to the Nodes module + refactor of tests to reflect the change.