fix(connectors): remove ets table; handle connector restart
Replace ETS table with a Registry; the Registry contains the function->supervisor relationships, where each function has a dedicated DynamicSupervisor. This supervisor then handles all Event Connector processes for the function; when the function is disconnected, the Supervisor is simply stopped. This fixes the issue where a process would restart, and the new pid would not be inserted in the ETS table (leading to an ineffective disconnect, where the old pid would be terminated instead of the new one).